+++ presentation/src/main/java/com/sms/presentation/main/ui/fill_out_information/component/ProjectsComponent.kt
@@ -9,24 +9,29 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.ModalBottomSheetState
+import androidx.compose.material.Text
 import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
+import com.msg.sms.design.component.picker.SmsDatePicker
 import com.msg.sms.design.component.toggle.ToggleComponent
+import com.msg.sms.design.modifier.smsClickable
+import com.msg.sms.design.theme.SMSTheme
 import com.sms.presentation.main.ui.fill_out_information.data.ProjectInfo
 import com.sms.presentation.main.ui.util.getFileNameFromUri
 import com.sms.presentation.main.ui.util.isImageExtensionCorrect
-import com.sms.presentation.main.viewmodel.FillOutViewModel
+import kotlinx.coroutines.launch
 
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
 fun ProjectsComponent(
     navController: NavController,
-    viewModel: FillOutViewModel,
-    detailStack: List<String>,
     bottomSheetState: ModalBottomSheetState,
+    detailStack: List<String>,
     data: ProjectInfo,
     savedData: (
         name: String,
@@ -37,10 +42,12 @@ fun ProjectsComponent(
         endDate: String,
         relatedLink: List<Pair<String, String>>
     ) -> Unit,
+    bottomSheetContent: @Composable (content: @Composable ColumnScope.() -> Unit) -> Unit,
     isImageExtensionInCorrect: (Boolean) -> Unit,
     onCancelButtonClick: () -> Unit
 ) {
     val context = LocalContext.current
+    val coroutineScope = rememberCoroutineScope()
     val projectName = remember {
         mutableStateOf(data.name)
     }
@@ -67,6 +74,9 @@ fun ProjectsComponent(
     }
     val isProjectProgress = remember {
         mutableStateOf(false)
+    }
+    val isProjectStartDate = remember {
+        mutableStateOf(true)
     }
     val singleSelectGalleryLauncher =
         rememberLauncherForActivityResult(ActivityResultContracts.PickVisualMedia()) { uri ->
@@ -121,9 +131,54 @@ fun ProjectsComponent(
         projectPreviewUriList,
         projectKeyTask.value,
         projectStartDate.value,
-        projectEndDate.value,
+        if (!isProjectProgress.value) projectEndDate.value else "",
         projectRelatedLinkList
     )
+
+    bottomSheetContent(content = {
+        val year = remember {
+            mutableStateOf(0)
+        }
+        val month = remember {
+            mutableStateOf(0)
+        }
+
+        SMSTheme { colors, typography ->
+            Box(
+                Modifier
+                    .fillMaxWidth()
+                    .padding(20.dp)
+            ) {
+                Text(
+                    text = "날짜 선택",
+                    style = typography.title2,
+                    color = colors.BLACK,
+                    fontWeight = FontWeight.Bold
+                )
+                Text(
+                    text = "완료",
+                    style = typography.body2,
+                    color = colors.P2,
+                    fontWeight = FontWeight.Normal,
+                    modifier = Modifier
+                        .align(Alignment.CenterEnd)
+                        .smsClickable {
+                            coroutineScope.launch { bottomSheetState.hide() }
+                            if (isProjectStartDate.value)
+                                projectStartDate.value = "${year.value}.${month.value}"
+                            else projectEndDate.value = "${year.value}.${month.value}"
+                        }
+                )
+            }
+            Spacer(modifier = Modifier.height(16.dp))
+            SmsDatePicker(yearValue = year.value,
+                monthValue = month.value,
+                yearRange = 2000..2030,
+                monthRange = 1..12,
+                onYearValueChange = { year.value = it },
+                onMonthValueChange = { month.value = it })
+        }
+    })
 
     ToggleComponent(name = "프로젝트", onCancelButtonClick = onCancelButtonClick) {
         Spacer(modifier = Modifier.height(32.dp))
@@ -163,8 +218,12 @@ fun ProjectsComponent(
             endDateText = projectEndDate.value,
             isProjectProgress = isProjectProgress.value,
             onStartDateCalendarClick = {
+                isProjectStartDate.value = true
+                coroutineScope.launch { bottomSheetState.show() }
             },
             onEndDateCalendarClick = {
+                isProjectStartDate.value = false
+                coroutineScope.launch { bottomSheetState.show() }
             },
             onProgressButtonClick = {
                 isProjectProgress.value = !isProjectProgress.value
