+++ presentation/src/main/java/com/sms/presentation/main/ui/fill_out_information/screen/ProjectsScreen.kt
@@ -1,7 +1,6 @@
 package com.sms.presentation.main.ui.fill_out_information.screen
 
 import android.net.Uri
-import android.util.Log
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
@@ -17,13 +16,11 @@ import androidx.navigation.NavController
 import com.msg.sms.design.component.SmsDialog
 import com.msg.sms.design.component.button.ButtonState
 import com.msg.sms.design.component.button.SmsRoundedButton
-import com.msg.sms.design.component.picker.SmsDatePicker
 import com.msg.sms.design.modifier.smsClickable
 import com.msg.sms.design.theme.SMSTheme
 import com.sms.presentation.main.ui.fill_out_information.component.*
 import com.sms.presentation.main.ui.fill_out_information.data.ProjectInfo
 import com.sms.presentation.main.viewmodel.FillOutViewModel
-import kotlinx.coroutines.launch
 
 @OptIn(ExperimentalMaterialApi::class)
 @Composable
@@ -38,12 +35,9 @@ fun ProjectsScreen(
     val projectList = remember {
         mutableStateListOf(*data.projects.toTypedArray())
     }
-    val coroutineScope = rememberCoroutineScope()
     val isImageExtensionInCorrect = remember {
         mutableStateOf(false)
     }
-
-    Log.d("ddd", data.projects.joinToString())
 
     if (isImageExtensionInCorrect.value) {
         SmsDialog(widthPercent = 1f,
@@ -55,61 +49,11 @@ fun ProjectsScreen(
             normalButtonOnClick = { isImageExtensionInCorrect.value = false })
     }
 
-    bottomSheetContent(content = {
-        val year = remember {
-            mutableStateOf(0)
-        }
-        val month = remember {
-            mutableStateOf(0)
-        }
-        val yearRange = remember {
-            mutableStateOf(2000..2030)
-        }
-        val monthRange = remember {
-            mutableStateOf(1..12)
-        }
-
-        SMSTheme { colors, typography ->
-            Box(
-                Modifier
-                    .fillMaxWidth()
-                    .padding(20.dp)
-            ) {
-                Text(
-                    text = "날짜 선택",
-                    style = typography.title2,
-                    color = colors.BLACK,
-                    fontWeight = FontWeight.Bold
-                )
-                Text(text = "완료",
-                    style = typography.body2,
-                    color = colors.P2,
-                    fontWeight = FontWeight.Normal,
-                    modifier = Modifier
-                        .align(Alignment.CenterEnd)
-                        .smsClickable {
-//                                if (isProjectStartDate.value) projectStartDate.value =
-//                                    "${year.value}.${month.value}"
-//                                else projectEndDate.value =
-//                                    "${year.value}.${month.value}"
-                            coroutineScope.launch { bottomSheetState.hide() }
-                        })
-            }
-            Spacer(modifier = Modifier.height(16.dp))
-            SmsDatePicker(yearValue = year.value,
-                monthValue = month.value,
-                yearRange = yearRange.value,
-                monthRange = monthRange.value,
-                onYearValueChange = { year.value = it },
-                onMonthValueChange = { month.value = it })
-        }
-    })
     LazyColumn {
         itemsIndexed(projectList) { idx, item ->
-            ProjectsComponent(navController = navController,
-                viewModel = viewModel,
+            ProjectsComponent(
+                navController = navController,
                 detailStack = detailStack,
-                bottomSheetState = bottomSheetState,
                 data = item,
                 savedData = { name, icon, preview, keyTask, startDate, endDate, relatedLink ->
                     projectList[idx] = ProjectInfo(
@@ -123,12 +67,15 @@ fun ProjectsScreen(
                         technologyOfUse = detailStack
                     )
                 },
+                bottomSheetState = bottomSheetState,
+                bottomSheetContent = bottomSheetContent,
                 isImageExtensionInCorrect = {
                     isImageExtensionInCorrect.value = it
                 },
                 onCancelButtonClick = {
                     projectList.removeAt(idx)
-                })
+                }
+            )
         }
         item {
             SMSTheme { colors, typography ->
